pyproject.toml
[project]
name = "artificial-mind-proactive-agent"
version = "0.1.0"
description = "This AI Agent helps user with their telecom billing, also it suggests plans according to user's requirement."
authors = [
    {name = "Tushar Sabale", email = "tushar.sabale@capgemini.com"}
    
]
readme = "README.md"
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.12"
requests = "^2.31.0"
google-adk = "^1.0.0"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.93.0"}
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"
deprecated = "^1.2.18"
jira = ">= 1.0.0"
twilio = "^9.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
